Selected TileElement now stored in paint session
